+++ utils/__init__.py
@@ -2,7 +2,4 @@
 """
 
 from . import elabConnector
-from . import IrodsConnectorAnonymous
-from . import IrodsConnectorIcommands
-from . import IrodsConnector
 from . import utils
